requirements.txt
fastapi
pymongo
motor
pydantic[email]
python-jose
passlib[bcrypt]
